Separate (blog-)posts and events into different sections
Unfortunately it seems that i18n for sections is not fully implemented
in hugo. Even when specifying the url in e.g.
'content/de/events/_index.md' it does not change the fact that the
individual pages later have the English slug in the URL (e.g.
'/events/...' instead of '/termine/...').

It's possible to work around this by using permalinks (configured in
config.toml), but this only works for top-level sections, not nested
ones (which fortunately we currently don't have).

The downside of the permalinks is that by default the filename is used
in the URL, but the slug is preferred if there is one. For the
translated content now I need to choose whether I want the :slug or the
:filename in the URL, without having this fallback mechanism. When the
:slug is missing it does fall back to the title (modified for the URL),
which seems like a reasonable compromise.

Some more info about the i18n section issue:

* https://discourse.gohugo.io/t/override-section-url/6197/13
* https://discourse.gohugo.io/t/how-to-translate-sections/9060/5
* https://github.com/gohugoio/hugo/issues/3354