Add timezone info to ics events
Found here: https://de.wikipedia.org/wiki/ICalendar

What is a little unclear to me is whether I should specify the timezone in the
date string (e.g. a Z at the end) or not and whether those time stamps should
actually be in UTC or not.

Also DTSTAMP has not TZID. Why?