feat: add script to parse and convert feedback and statistics
The original files are not stored in this repository for privacy reasons. This
script transforms that original data into data that can then be used for
further analysis and plotting.

Extract things from the existing script into this one where it makes sense.

Note that parse.py does not generate all data/ files as the previous
generate.py did, for this reason those files are kept around for now. In future
changes generate.py will undergo major changes, therefore I don't want to fix
it now to use the new files.